+++ tests/src/lib.rs
@@ -1,2 +1,2 @@
 #[cfg(test)]
-mod connections;
+mod tests;
